a little clean up and file renaming.
